move publicationInfo lcpInfo part to dedicated file
