+++ src/renderer/components/dialog/publicationInfos/PublicationInfo.tsx
@@ -24,11 +24,10 @@ import Loader from "readium-desktop/renderer/components/utils/Loader";
 import { RootState } from "readium-desktop/renderer/redux/states";
 import { TDispatch } from "readium-desktop/typings/redux";
 
-import { StatusEnum } from "@r2-lcp-js/parser/epub/lsd";
-
 import Dialog from "../Dialog";
 import CatalogControls from "./catalogControls";
 import CatalogLcpControls from "./catalogLcpControls";
+import LcpInfo from "./lcpInfo";
 import OpdsControls from "./opdsControls";
 
 // tslint:disable-next-line: no-empty-interface
@@ -93,69 +92,6 @@ class PublicationInfo extends React.Component<IProps, IState> {
             publication.publishers.length
             ? publication.publishers.join(", ")
             : undefined;
-
-        const locale = this.props.translator.getLocale();
-
-        // https://momentjs.com/docs/#/displaying/
-        moment.locale(locale);
-
-        const lcp = publication.lcp;
-        const lcpRightsPrint = (lcp?.rights?.print) ? lcp.rights.print : 0;
-        const lcpRightsCopy = (lcp?.rights?.copy) ? lcp.rights.copy : 0;
-        const lcpRightsCopies = publication.lcpRightsCopies ?? 0;
-
-        const lcpRightsStartDate = (lcp?.rights?.start) ? lcp.rights.start : undefined;
-        let lcpRightsStartDateStr: string | undefined;
-        if (lcpRightsStartDate) {
-            try {
-                lcpRightsStartDateStr = moment(lcpRightsStartDate).format("LLL");
-            } catch (err) {
-                console.log(err);
-                try {
-                    lcpRightsStartDateStr = lcpRightsStartDate.toLocaleString(locale);
-                } catch (err2) {
-                    console.log(err2);
-                    lcpRightsStartDateStr = lcpRightsStartDate.toLocaleString();
-                }
-            }
-        }
-
-        const lcpRightsEndDate = (lcp?.rights?.end) ? lcp.rights.end : undefined;
-        let lcpRightsEndDateStr: string | undefined;
-        if (lcpRightsEndDate) {
-            try {
-                lcpRightsEndDateStr = moment(lcpRightsEndDate).format("LLL");
-            } catch (err) {
-                console.log(err);
-                try {
-                    lcpRightsEndDateStr = lcpRightsEndDate.toLocaleString(locale);
-                } catch (err2) {
-                    console.log(err2);
-                    lcpRightsEndDateStr = lcpRightsEndDate.toLocaleString();
-                }
-            }
-        }
-
-        // TODO: fix r2-lcp-js to handle encrypted fields
-        // (need lcp.node with userkey decrypt, not contentkey):
-        // if (lcp && lcp.r2LCPBase64) {
-        //     const r2LCPStr = Buffer.from(lcp.r2LCPBase64, "base64").toString("utf-8");
-        //     const r2LCPJson = JSON.parse(r2LCPStr);
-        //     const r2LCP = TaJsonDeserialize<LCP>(r2LCPJson, LCP);
-        //     r2LCP.JsonSource = r2LCPStr;
-
-        //     console.log(r2LCP.User.Name);
-        //     console.log(r2LCP.User.Email);
-        //     console.log(JSON.stringify(r2LCP.User.Encrypted, null, 4));
-        // }
-
-        const lsdOkay = lcp &&
-            lcp.lsd &&
-            lcp.lsd.lsdStatus;
-
-        const lsdStatus = lsdOkay &&
-            lcp.lsd.lsdStatus.status ?
-            lcp.lsd.lsdStatus.status : undefined;
 
         const renderInfo = () =>
             <>
@@ -228,46 +164,8 @@ class PublicationInfo extends React.Component<IProps, IState> {
                             <br />
                         </p>
 
-                        {lcp && <>
-                            <h3>LCP</h3>
-                            <p className={classNames(styles.allowUserSelect)}
-                            >
-                                {(lsdStatus &&
-                                    (lsdStatus !== StatusEnum.Active && lsdStatus !== StatusEnum.Ready)) && <>
-                                <span style={{color: "red"}}>{(lsdStatus === StatusEnum.Expired ?
-                                        __("publication.expiredLcp")
-                                        : ((lsdStatus === StatusEnum.Cancelled) ?
-                                        __("publication.cancelledLcp")
-                                        : ((lsdStatus === StatusEnum.Revoked) ?
-                                        __("publication.revokedLcp")
-                                        : (lsdStatus === StatusEnum.Returned ?
-                                        __("publication.returnedLcp") :
-                                        `LCP LSD: ${lsdStatus}`))))}</span>
-                                <br /><br />
-                                </>}
+                        <LcpInfo publicationLcp={publication}></LcpInfo>
 
-                                {lcpRightsStartDateStr && <>
-                                <span>{__("publication.lcpStart")}: </span><i>{lcpRightsStartDateStr}</i>
-                                <br />
-                                </>}
-
-                                {lcpRightsEndDateStr && <>
-                                <span>{__("publication.lcpEnd")}: </span><i>{lcpRightsEndDateStr}</i>
-                                <br />
-                                <br />
-                                </>}
-
-                                {lcpRightsCopy && <>
-                                <span>{__("publication.lcpRightsCopy")}: </span>
-                                <i>{lcpRightsCopies} / {lcpRightsCopy}</i><br />
-                                </>}
-
-                                {lcpRightsPrint && <>
-                                <span>{__("publication.lcpRightsPrint")}: </span>
-                                <i>0 / {lcpRightsPrint}</i><br />
-                                </>}
-                            </p>
-                        </>}
                     </div>
                 </div>
             </>;
